% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.sbc_build_connection_params}
\alias{.sbc_build_connection_params}
\title{Construir parámetros de conexión según tipo de BD}
\usage{
.sbc_build_connection_params(tipo, config)
}
\arguments{
\item{tipo}{Tipo de base de datos ("mysql" o "db2")}

\item{config}{Lista con configuración de conexión}
}
\value{
Lista de parámetros para DBI::dbConnect()
}
\description{
Construir parámetros de conexión según tipo de BD
}
\keyword{internal}
