% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.sbc_get_driver}
\alias{.sbc_get_driver}
\title{Obtener driver apropiado según tipo de base de datos}
\usage{
.sbc_get_driver(tipo, driver_path = NULL)
}
\arguments{
\item{tipo}{Tipo de base de datos: "mysql" o "db2"}

\item{driver_path}{Ruta al driver JDBC (solo para DB2)}
}
\value{
Objeto driver DBI
}
\description{
Obtener driver apropiado según tipo de base de datos
}
\keyword{internal}
