% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.sbc_validate_config}
\alias{.sbc_validate_config}
\title{Validar configuración de conexión}
\usage{
.sbc_validate_config(config, tipo)
}
\arguments{
\item{config}{Lista con parámetros de conexión}

\item{tipo}{Tipo de base de datos}
}
\value{
TRUE si válida, lanza error en caso contrario
}
\description{
Validar configuración de conexión
}
\keyword{internal}
