% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sbc_get_db2_config}
\alias{sbc_get_db2_config}
\title{Cargar configuración de base de datos DB2}
\usage{
sbc_get_db2_config(config_name = "DB_CONFIG_DB2")
}
\arguments{
\item{config_name}{Nombre del archivo de configuración (sin extensión)}
}
\value{
Lista con la configuración de la base de datos
}
\description{
Lee el archivo de configuración de base de datos DB2 guardado previamente
}
\examples{
# Cargar configuración de DB2
db_config <- sbc_get_db2_config()

# Cargar configuración con nombre personalizado
db_config <- sbc_get_db2_config("DB_CONFIG_DB2")
}
