% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sbc_get_mysql_config}
\alias{sbc_get_mysql_config}
\title{Cargar configuración de base de datos MySQL}
\usage{
sbc_get_mysql_config(config_name = "DB_CONFIG_MYSQL")
}
\arguments{
\item{config_name}{Nombre del archivo de configuración (sin extensión)}
}
\value{
Lista con la configuración de la base de datos
}
\description{
Lee el archivo de configuración de base de datos MySQL guardado previamente
}
\examples{
# Cargar configuración de MySQL
db_config <- sbc_get_mysql_config()

# Cargar configuración con nombre personalizado
db_config <- sbc_get_mysql_config("DB_CONFIG_MYSQL")
}
